more cleanup, docs
